requirements.txt
cvxopt
scikit-learn
numpy
torch
anndata
matplotlib
absl-py
scipy
mofapy
mygene